// js/src/types/Fund.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BudgetItem } from "./Budget";

export type BudgetFund = { id: number, name: string, };

export type BudgetFundFields = { name: string, };

export type GetAllFundsResponse = { funds: Array<BudgetFund>, items: Array<BudgetItem>, };